Development/software/requirements.txt
numpy
pygame
sounddevice
soundfile
openpyxl
Pillow
